Lista la lógica de la estrategia A, sin llenar la tabla

+++ Montecarlo/Montecarlo/Acumuladores/IAcumulador.cs
@@ -2,6 +2,6 @@
 {
     public interface IAcumulador
     {
-        decimal CalcularSiguiente(decimal valor);
+        double CalcularSiguiente(double valor);
     }
 }
